Dev : Refactor folders and file names
