+++ src/components/index.ts
@@ -1,6 +1,12 @@
-export * from './add-balance-form';
-export * from './card-balances';
-export * from './header';
+export * from './auth/AnonymousSignInButton';
+export * from './auth/ConfirmEmail';
+export * from './auth/LoginForm';
+export * from './auth/ResetPasswordForm';
+export * from './auth/SignUpForm';
+export * from './dashboard/CardBalances';
+export * from './dashboard/DashboardLayout';
+export * from './dashboard/Header';
+export * from './dashboard/Home';
 export * from './ui/badge';
 export * from './ui/breadcrumb';
 export * from './ui/button-loading';
