+++ src/components/ui/breadcrumb.tsx
@@ -2,7 +2,7 @@ import * as React from "react"
 import { ChevronRightIcon, DotsHorizontalIcon } from "@radix-ui/react-icons"
 import { Slot } from "@radix-ui/react-slot"
 
-import { cn } from "@/lib/cnutils"
+import { cn } from "@/utils/cnutils"
 
 const Breadcrumb = React.forwardRef<
   HTMLElement,
