+++ src/components/ui/dropdown-menu.tsx
@@ -6,7 +6,7 @@ import {
   DotFilledIcon,
 } from "@radix-ui/react-icons"
 
-import { cn } from "@/lib/cnutils"
+import { cn } from "@/utils/cnutils"
 
 const DropdownMenu = DropdownMenuPrimitive.Root
