+++ src/components/ui/form.tsx
@@ -10,7 +10,7 @@ import {
   useFormContext,
 } from "react-hook-form"
 
-import { cn } from "@/lib/cnutils"
+import { cn } from "@/utils/cnutils"
 import { Label } from "@/components/ui/label"
 
 const Form = FormProvider
