+++ src/components/ui/password-input.tsx
@@ -2,7 +2,7 @@ import { forwardRef, useState } from "react";
 import { EyeIcon, EyeOffIcon } from "lucide-react";
 import { Button } from "@/components/ui/button";
 import { Input, InputProps } from "@/components/ui/input";
-import { cn } from "@/lib/cnutils";
+import { cn } from "@/utils/cnutils";
 
 const PasswordInput = forwardRef<HTMLInputElement, InputProps>(
   ({ className, ...props }, ref) => {
