+++ src/components/ui/toast.tsx
@@ -3,7 +3,7 @@ import { Cross2Icon } from "@radix-ui/react-icons"
 import * as ToastPrimitives from "@radix-ui/react-toast"
 import { cva, type VariantProps } from "class-variance-authority"
 
-import { cn } from "@/lib/cnutils"
+import { cn } from "@/utils/cnutils"
 
 const ToastProvider = ToastPrimitives.Provider
