+++ src/lib/index.ts
@@ -1,18 +1,4 @@
-export * from './api/getLatestBalance';
-export * from './api/getUser';
-export * from './api/listenAuthEvents';
-export * from './api/loginUser';
-export * from './api/logoutUser';
-export * from './api/postBalance';
-export * from './api/signUpUser';
-export * from './cnutils';
-export * from './constants/AmountSchema';
-export * from './constants/expenseSchema';
-export * from './emails/confirmation-email';
-export * from './emails/send-email-confirmation';
-export * from './functions/enum';
-export * from './types/Balance';
-export * from './types/Expenses';
-export * from './types/LoginUserFormSchema';
-export * from './types/SignUpFormSchema';
-export * from './types/supabase-generated.d';
+export * from './schemas/amountSchema';
+export * from './schemas/expenseSchema';
+export * from './schemas/loginUserFormSchema';
+export * from './schemas/signUpFormSchema';
