--- src/pages/index.ts
@@ -1,6 +1,0 @@
-export * from './auth/anonymous-sign-in-button';
-export * from './auth/confirm-email';
-export * from './auth/login-form';
-export * from './auth/reset-passwordForm';
-export * from './auth/sign-up-form';
-export * from './dashboard-home';
