+++ src/routes/router.tsx
@@ -1,28 +1,12 @@
+import { LoginForm, SignUpForm } from "@/components";
+import ConfirmEmail from "@/components/auth/ConfirmEmail";
+import ResetPasswordForm from "@/components/auth/ResetPasswordForm";
+import DashboardLayout from "@/components/dashboard/DashboardLayout";
+import DashboardHome from "@/components/dashboard/Home";
+import { isAuthenticated } from "@/hooks/auth/isAuth";
+import ProtectedRoute from "@/services/auth/ProtectedRoute";
 import { createBrowserRouter } from "react-router-dom";
-import { LoginForm, SignUpForm } from "@/pages";
-import { getSession } from "@/lib/api/getSession.ts";
-import DashboardHome from "@/pages/dashboard-home.tsx";
-import ProtectedRoute from "../Protected-route";
-import DashboardLayout from "@/layout/dashboard-layout";
-import ConfirmEmail from "@/pages/auth/confirm-email";
-import ResetPasswordForm from "@/pages/auth/reset-passwordForm";
 
-const getAccessToken = async (): Promise<string | null> => {
-  const session = await getSession();
-  console.log("access_token", session?.access_token);
-  if (!session?.access_token) {
-    return null;
-  }
-  return session.access_token;
-};
-
-const isAuthenticated = async (): Promise<boolean | null> => {
-  const accessToken = await getAccessToken();
-  if (!accessToken) {
-    return false;
-  }
-  return true;
-};
 
 export const router = createBrowserRouter([
   {
