+++ src/services/emails/SendEmailConfirmation.tsx
@@ -1,5 +1,5 @@
 import { Resend } from "resend";
-import ConfirmationEmail from "./confirmation-email";
+import ConfirmationEmail from "./ConfirmatiomEmail";
 
 const resend = new Resend(import.meta.env.VITE_RESEND_API_KEY);
