+++ src/types/supabase-generated.d.ts
